requirements.txt
nfstream
pandas
mysql-connector-python
requests
python-dotenv
schedule
psutil
loguru
